pyproject.toml
[tool.poetry]
name = "toy-fastapi-app"
version = "0.1.0"
description = ""
authors = ["shoang22 <54875725+shoang22@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
langchain = "^0.1.16"
uvicorn = "^0.29.0"
pydantic = "^2.7.0"
taskiq-fastapi = "^0.3.1"
taskiq-nats = "^0.4.0"
taskiq-redis = "^0.5.5"
gunicorn = "^22.0.0"
locust = "^2.26.0"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
fastapi-profiler = "^1.3.0"
taskiq = {extras = ["reload"], version = "^0.11.3"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
